Update code examples with cancellation functions.
Introduce a new section "code common to examples in this paper" presenting
assert_cancel(), a cancellation function that asserts if it is ever reached,
and launch(), a factory function that provides an entry-function wrapper and a
cancellation function and returns a new fiber_context instance.

Update code examples to either pass assert_cancel to fiber_context constructor
or use launch() factory function.