Update 'rev' script to collect 'Changes since' rather than replacing.
It seems that it might be useful to record some of the back-and-forth changes
to the P0876 series of papers. So instead of changing an existing 'Changes
since $prevrev' line (implying replacing the text that follows), insert a new
'Changes since $oldrev' above 'Changes since $prevrev', leaving a section to
be filled in.

Name the backup files from the sed commands .tex~ instead of .tex.tex~.