Accept span<byte, N> instead of void* as explicit stack area.
Even though the fiber_context library has no checks to constrain the stack to
the specified size, it's still important to specify both ends of the passed
area, since portable code can't know in which direction the running CPU will
consume the stack. If fiber_context() accepted only a void* pointer, which end
of the stack block should the caller pass?